backend/pyproject.toml
[project]
name = "temp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[standard]>=0.118.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]
